complete

+++ Develop/routes/api/product-routes.js
@@ -9,7 +9,7 @@ router.get('/', async (req, res) => {
     const productData = await Product.findAll({
       include: [{ model: Category }]
     }, {
-      include: [{ model: Tag, through: ProductTag, as: 'products_with_tags'}]
+      include: [{ model: Tag, through: ProductTag, as: 'products_with_tags' }]
     });
     res.status(200).json(productData);
   } catch (err) {
@@ -21,18 +21,18 @@ router.get('/', async (req, res) => {
 router.get('/:id', async (req, res) => {
   try {
     const productData = await Product.findByPk(req.params.id, {
-      include: [{ model: Category}]
+      include: [{ model: Category }]
     }, {
       include: [{ model: Tag }]
     })
 
-    if(!productData){
+    if (!productData) {
       res.status(404).json({ message: 'No product with this id!' });
       return;
     }
 
     res.status(200).json(productData);
-  } catch(err) {
+  } catch (err) {
     res.status(500).json(err);
   }
 });
@@ -49,8 +49,9 @@ router.post('/', (req, res) => {
   */
   Product.create(req.body)
     .then((product) => {
+      //if no tagids --> create error message
       // if there's product tags, we need to create pairings to bulk create in the ProductTag model
-      if (req.body.tagIds.length) {
+      if (req.body.tagIds?.length) {
         const productTagIdArr = req.body.tagIds.map((tag_id) => {
           return {
             product_id: product.id,
@@ -69,7 +70,6 @@ router.post('/', (req, res) => {
     });
 });
 
-//! Need help .. not sure ... working? how to test
 router.put('/:id', (req, res) => {
   // update product data
   Product.update(req.body, {
@@ -78,31 +78,37 @@ router.put('/:id', (req, res) => {
     }
   })
     .then((product) => {
+      // if no tagIds in req.body --> skip modification of tagids
+      if (!req.body.tagIds) {
+        return product;
+      }
       // find all associated tags from ProductTag
-      return ProductTag.findAll({ where: { product_id: req.params.id } });
-    })
-    .then((productTags) => {
-      // get list of current tag_ids
-      const productTagIds = productTags.map(({ tag_id }) => tag_id);
-      // create filtered list of new tag_ids
-      const newProductTags = req.body.tagIds
-        .filter((tag_id) => !productTagIds.includes(tag_id))
-        .map((tag_id) => {
-          return {
-            product_id: req.params.id,
-            tag_id,
-          };
-        });
-      // figure out which ones to remove
-      const productTagsToRemove = productTags
-        .filter(({ tag_id }) => !req.body.tagIds.includes(tag_id))
-        .map(({ id }) => id);
+      return ProductTag.findAll({ where: { product_id: req.params.id } })
 
-      // run both actions
-      return Promise.all([
-        ProductTag.destroy({ where: { id: productTagsToRemove } }),
-        ProductTag.bulkCreate(newProductTags),
-      ]);
+        .then((productTags) => {
+          
+          // get list of current tag_ids
+          const productTagIds = productTags.map(({ tag_id }) => tag_id);
+          // create filtered list of new tag_ids
+          const newProductTags = req.body.tagIds
+            .filter((tag_id) => !productTagIds.includes(tag_id))
+            .map((tag_id) => {
+              return {
+                product_id: req.params.id,
+                tag_id,
+              };
+            });
+          // figure out which ones to remove
+          const productTagsToRemove = productTags
+            .filter(({ tag_id }) => !req.body.tagIds.includes(tag_id))
+            .map(({ id }) => id);
+
+          // run both actions
+          return Promise.all([
+            ProductTag.destroy({ where: { id: productTagsToRemove } }),
+            ProductTag.bulkCreate(newProductTags),
+          ]);
+        })
     })
     .then((updatedProductTags) => res.json(updatedProductTags))
     .catch((err) => {
@@ -111,7 +117,6 @@ router.put('/:id', (req, res) => {
     });
 });
 
-//!
 router.delete('/:id', async (req, res) => {
   // delete one product by its `id` value
   try {
@@ -127,7 +132,7 @@ router.delete('/:id', async (req, res) => {
     }
 
     res.status(200).json(productData);
-  } catch(err) {
+  } catch (err) {
     res.status(500).json(err);
   }
 });
